refactoring(problem7): 입력값 검증 클래스를 입력값 단위의 검증 클래스로 각각 분리
