+++ src/main/java/onboarding/problem7/collection/Visitors.java
@@ -1,6 +1,6 @@
 package onboarding.problem7.collection;
 
-import onboarding.problem7.validation.FriendValidator;
+import onboarding.problem7.validation.VisitorsValidator;
 
 import java.util.List;
 import java.util.function.Consumer;
@@ -13,7 +13,7 @@ public class Visitors {
     }
 
     public static Visitors of(List<String> visitors) {
-        FriendValidator.validateVisitors(visitors);
+        VisitorsValidator.validateVisitors(visitors);
         return new Visitors(visitors);
     }
